refactor printing
