+++ routes/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod contacts;
 pub mod dictionary;
 pub mod faqs;
+pub mod printing;
