+++ src/routes.rs
@@ -1,5 +1,4 @@
 pub mod food;
-pub mod printing;
 pub mod spaces;
 pub mod tools;
 pub mod transit;
